agents/pyproject.toml
[tool.poetry]
name = "agents"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
PyJWT = "^2.8.0"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
langchain-groq = "^0.2.0"
langchain-community = "^0.3.3"
langgraph = "^0.2.39"
numexpr = "^2.10.1"
wikipedia-api = "^0.7.1"
wikipedia = "^1.4.0"
cohere = "^5.11.3"
langsmith = "^0.1.142"
langchain-chroma = "^0.1.4"
langchain-google-community = "^2.0.2"
google-api-python-client = "^2.151.0"
html2text = "^2024.2.26"
grandalf = "^0.8"
wolframalpha = "^5.1.3"
yfinance = "^0.2.50"
langfuse = "^2.54.1"
pydantic-settings = "^2.6.1"
json-repair = "^0.30.1"
psycopg2-binary = "^2.9.10"
finnhub-python = "^2.4.21"
alpha-vantage = "^3.0.0"
python-multipart = "^0.0.17"
pypdf2 = "^3.0.1"
pymupdf = "^1.24.14"
altair = "^5.5.0"
plotly = "^5.24.1"
pdfkit = "^1.0.0"
markdown2 = "^2.5.1"
tavily-python = "^0.5.0"
psutil = "^6.1.0"
weasyprint = "63.0"
vl-convert-python = "^1.7.0"
langchain-openai = "^0.2.10"
markdown = "^3.5.1"
selenium = "^4.27.1"
pyjwt = "^2.8.0"
kaleido = "0.2.1"
pytesseract = "^0.3.13"
diskcache = "^5.6.3"
chromedriver-autoinstaller = "^0.6.4"

[[tool.poetry.source]]
name = "pytorch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
